SPARK-13688: Add spark.dynamicAllocation.overrideNumInstances.
When both the number of executor instances and dynamic allocation are
configured, dynamic allocation is ignored and the number of executors is
statically set. This introduces a configuration property to change this
behavior so that cluster administrators can make dynamic allocation the
default when users set --num-executors instead of
spark.dynamicAllocation.minExecutors and unintentionally disable dynamic
allocation for a job.