NETFLIX-BUILD: Reducing the total number of history server replay threads to 25% of available cores.
By default history server runs with number of replay threads = number of available cores. This can result in history server starving other processes running on the same host , specially on statup when it tries to process all the old files it ends up consuming all the CPU. This change limits number of cores used to 25% of available cores to avoid starvation of other processes.