NETFLIX-BUILD: Do not preload file listings for all partitions.
When converting a MetastoreRelation to a ParquetRelation, Hive was
passing all partitions, which HadoopFsRelation (superclass of
ParquetRelation) was listing each directory and caching the result to
speed up queries. The problem was that this happens before predicates
are known and Spark lists all table partitions. This change updates the
conversion to pass an empty list of partitions so there are no paths
pre-cached. It also detects what partitions are missing and adds them to
the cache just before returning results so that FileStatuses are cached,
but only for partitions that are needed for queries.